대시보드/헤드라인 컴포넌트화
